+++ R/resp_val.R
@@ -2,5 +2,6 @@
 #' @importFrom insight get_response
 #' @export
 resp_val <- function(x) {
+  .Deprecated("insight::get_response()")
   insight::get_response(x)
 }