+++ R/resp_var.R
@@ -2,5 +2,6 @@
 #' @importFrom insight find_response
 #' @export
 resp_var <- function(x, combine = TRUE) {
+  .Deprecated("insight::find_response()")
   insight::find_response(x = x, combine = combine)
 }